% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/attritTAPS.R
\docType{methods}
\name{attritTAPS}
\alias{attritTAPS}
\alias{attritTAPS,ANY-method}
\title{Find attrition stats for TAPS waves}
\usage{
attritTAPS(outcome)
}
\arguments{
\item{outcome}{A character vector of the names of outcome variables of interest}
}
\value{
A list of attrition rates in population margins. The number in each category corresponds to the percentage of people within that category that left/attrited the panel
}
\description{
Find response attrition statistics for TAPS waves
}
\author{
David G. Carlson \email{carlson.david@wustl.edu} Michelle Torres: \email{smtorres@wustl} Taeyong Park \email{t.park@wustl.edu}
}
\seealso{
\code{\link{weightTAPSPACK}} \code{\link{weightTAPS}} \code{\link{variablesTAPS}} \code{\link{weightTAPSoutput}} \code{\link{simpleWeight}} \code{\link{subsetTAPS}} \code{\link{multipleImp}} \code{\link{hotdeckImp}} \code{\link{wavesTAPS}}
}

