\name{dd}
\docType{data}
\alias{dd}
\title{TAPS imputed demographics}
\description{
   This is the imputed demographics for the TAPS survey.
}
\usage{dd}
\format{This is a .rda file}
\seealso{
\code{\link{weightTAPSPACK}} \code{\link{weightTAPS}}
\code{\link{weightTAPSoutput}} \code{\link{simpleWeight}}
\code{\link{attritTAPS}} \code{\link{multipleImp}}
\code{\link{hotdeckImp}} \code{\link{subsetTAPS}} 
\code{\link{wavesTAPS}}
}
