% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hotdeckImp.R
\docType{methods}
\name{hotdeckImp}
\alias{hotdeckImp}
\alias{hotdeckImp,ANY-method}
\title{Impute data using hotdeck imputation method}
\usage{
hotdeckImp(data, outcome, covars)
}
\arguments{
\item{data}{A data frame with missing values to be imputed}

\item{outcome}{Character strings indicating outcome variable(s) of the user's interest}

\item{covars}{Character strings indicating covariates of the user's interest}
}
\value{
A data frame with imputed values
}
\description{
Impute missing data in a dataframe using hotdeck imputation method
}
\author{
David G. Carlson \email{carlson.david@wustl.edu}  Michelle Torres: \email{smtorres@wustl} Taeyong Park \email{t.park@wustl.edu}
}
\seealso{
\code{\link{weightTAPSPACK}} \code{\link{weightTAPS}} \code{\link{variablesTAPS}} \code{\link{weightTAPSoutput}} \code{\link{simpleWeight}} \code{\link{subsetTAPS}} \code{\link{attritTAPS}} \code{\link{hotdeckImp}} \code{\link{wavesTAPS}}
}

