% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multipleImp.R
\docType{methods}
\name{multipleImp}
\alias{multipleImp}
\alias{multipleImp,ANY-method}
\title{Impute data using multiple imputation method}
\usage{
multipleImp(data, outcome, covars, m = 5)
}
\arguments{
\item{data}{A data frame with missing values to be imputed}

\item{outcome}{Character strings indicating outcome variable(s) of the user's interest}

\item{covars}{Character strings indicating covariates of the user's interest}

\item{m}{A nunmeric object indicating number of data sets to create - default is 5}
}
\value{
A data frame with imputed values
}
\description{
Impute missing data in a dataframe using multiple imputation method
}
\author{
David G. Carlson \email{carlson.david@wustl.edu} Michelle Torres: \email{smtorres@wustl} Taeyong Park \email{t.park@wustl.edu}
}
\seealso{
\code{\link{weightTAPSPACK}} \code{\link{weightTAPS}} \code{\link{variablesTAPS}} \code{\link{weightTAPSoutput}} \code{\link{simpleWeight}} \code{\link{subsetTAPS}} \code{\link{attritTAPS}} \code{\link{hotdeckImp}} \code{\link{wavesTAPS}}
}

