\name{pop.margins}
\docType{data}
\alias{pop.margins}
\title{Population margins}
\description{
   Population margins for the U.S.
}
\usage{pop.margins}
\format{This is a .rda file}
\seealso{
\code{\link{weightTAPSPACK}} \code{\link{weightTAPS}}
\code{\link{weightTAPSoutput}} \code{\link{simpleWeight}}
\code{\link{attritTAPS}} \code{\link{multipleImp}}
\code{\link{hotdeckImp}} \code{\link{subsetTAPS}} 
\code{\link{wavesTAPS}}
}
