\docType{data}
\name{TAPSdata}
\alias{TAPSdata}
\title{TAPS data}
\description{
   This is the data for the TAPS survey. The codebook can be found at \url{http://taps.wustl.edu/data-archive}
}
\usage{TAPScum}
\format{This is a .rda file}
\source{\url{taps.wustl.edu}}
\seealso{
\code{\link{weightTAPSPACK}} \code{\link{weightTAPS}}
\code{\link{weightTAPSoutput}} \code{\link{simpleWeight}}
\code{\link{attritTAPS}} \code{\link{multipleImp}}
\code{\link{hotdeckImp}} \code{\link{subsetTAPS}} 
\code{\link{wavesTAPS}} \code{\link{TAPSimputeddemographics}}
}
