\name{TAPSimputeddemographics}
\docType{data}
\alias{TAPSimputeddemographics}
\title{TAPS demographics data}
\description{
   This is the data for the demographics of the TAPS survey.
}
\usage{dd}
\format{This is a .rda file and tab delineated}
\source{taps.wustl.edu}
\seealso{
\code{\link{weightTAPSPACK}} \code{\link{weightTAPS}}
\code{\link{weightTAPSoutput}} \code{\link{simpleWeight}}
\code{\link{attritTAPS}} \code{\link{multipleImp}}
\code{\link{hotdeckImp}} \code{\link{subsetTAPS}} 
\code{\link{wavesTAPS}}
}
