% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variablesTAPS.R
\docType{methods}
\name{variablesTAPS}
\alias{variablesTAPS}
\alias{variablesTAPS,ANY-method}
\title{Lists names of variables by wave}
\usage{
variablesTAPS(month, year)
}
\arguments{
\item{month}{A character vector specifying the first three letters of the month(s) of the wave(s) of interest}

\item{year}{A numeric vector specifying the year(s) of the wave(s) desired. This vector MUST BE the same length of the month vector (one year per month specified)}
}
\value{
A list with the names of the variables per wave specified
}
\description{
Lists the names of the variables in the TAPS dataset for the wave(s) specified
}
\examples{
variablesTAPS(month=c("Feb","Mar"), year=c(2012,2012))
}
\author{
David G. Carlson \email{carlson.david@wustl.edu}  Michelle Torres: \email{smtorres@wustl}
}
\seealso{
\code{\link{weightTAPS}} \code{\link{weightTAPSPACK}} \code{\link{subsetTAPS}} \code{\link{weightTAPSoutput}} \code{\link{simpleWeight}} \code{\link{attritTAPS}} \code{\link{multipleImp}} \code{\link{hotdeckImp}} \code{\link{wavesTAPS}}
}

