% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wavesTAPS.R
\docType{methods}
\name{wavesTAPS}
\alias{wavesTAPS}
\alias{wavesTAPS,ANY-method}
\title{Find population statistics for different waves}
\usage{
wavesTAPS(outcome)
}
\arguments{
\item{outcome}{A character vector of the names of outcome variables of interest}
}
\value{
A list of population stats
}
\description{
Find the population statistics for each wave in TAPS data to be analyzed. Useful for attrition information.
}
\author{
David G. Carlson \email{carlson.david@wustl.edu}  Michelle Torres: \email{smtorres@wustl} Taeyong Park \email{t.park@wustl.edu}
}
\seealso{
\code{\link{weightTAPSPACK}} \code{\link{weightTAPS}} \code{\link{variablesTAPS}} \code{\link{weightTAPSoutput}} \code{\link{simpleWeight}} \code{\link{attritTAPS}} \code{\link{multipleImp}} \code{\link{hotdeckImp}} \code{\link{subsetTAPS}}
}

